Add new User. RUN

+++ src/main/java/ru/alvisid/testtaskmt/UserRepository.java
@@ -1,5 +1,6 @@
 package ru.alvisid.testtaskmt;
 
 import org.springframework.data.repository.CrudRepository;
+import ru.alvisid.testtaskmt.model.User;
 
-public interface UserRepository extends CrudRepository <User, Long> {}
+public interface UserRepository extends CrudRepository <User, Integer> {}
